Modificar o programa para adicionar e excluir categoria só com o gerente e operacional

+++ src/main/java/br/com/zup/sistemareembolso/services/CategoriaService.java
@@ -1,9 +1,9 @@
 package br.com.zup.sistemareembolso.services;
 
-import br.com.zup.sistemareembolso.exceptions.CategoriaNaoExistenteException;
-import br.com.zup.sistemareembolso.exceptions.CategoriaRepetidaException;
-import br.com.zup.sistemareembolso.exceptions.ColaboradorNaoExistenteException;
+import br.com.zup.sistemareembolso.exceptions.*;
+import br.com.zup.sistemareembolso.models.Cargo;
 import br.com.zup.sistemareembolso.models.Categoria;
+import br.com.zup.sistemareembolso.models.Colaborador;
 import br.com.zup.sistemareembolso.repositories.CategoriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,7 +16,16 @@ public class CategoriaService {
     @Autowired
     private CategoriaRepository categoriaRepository;
 
-    public  Categoria  salvarCategoria ( Categoria  categoria ) {
+    @Autowired
+    private ColaboradorService colaboradorService;
+
+    public  Categoria  salvarCategoria ( Categoria  categoria, Colaborador colaborador) {
+        Colaborador colaboradorDoBanco = colaboradorService.pesquisarColaboradorPorCpf(colaborador.getCpf());
+
+        if (!colaboradorDoBanco.getCargo().equals(Cargo.GERENTE) && !colaboradorDoBanco.getCargo().equals(Cargo.DIRETOR)) {
+            throw new PermissaoNegadaParaCriarCategoriaException();
+        }
+
         if (categoriaRepository.existsByDescricao(categoria.getDescricao())) {
             throw new CategoriaRepetidaException();
         }
@@ -38,7 +47,13 @@ public class CategoriaService {
         throw new CategoriaNaoExistenteException();
     }
 
-    public  void  deletarCategoria(Integer codCategoria) {
+    public  void  deletarCategoria(Integer codCategoria, Colaborador colaborador) {
+        Colaborador colaboradorDoBanco = colaboradorService.pesquisarColaboradorPorCpf(colaborador.getCpf());
+
+        if (!colaboradorDoBanco.getCargo().equals(Cargo.GERENTE) && !colaboradorDoBanco.getCargo().equals(Cargo.DIRETOR)) {
+            throw new PermissaoNegadaParaExcluirCategoriaException();
+        }
+
         Categoria categoria = pesquisarCategoriaPorCodCategoria(codCategoria);
         categoriaRepository.delete(categoria);
     }
